requirements.txt
acres==0.3.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
etelemetry==0.3.1
filelock==3.18.0
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.10
isodate==0.6.1
Levenshtein==0.27.1
looseversion==1.3.0
lxml==5.3.1
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
prov==2.0.1
puremagic==1.28
pydot==3.0.4
PyMuPDF==1.25.5
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
python-telegram-bot==20.3
pytz==2025.2
pyxnat==1.6.3
RapidFuzz==3.13.0
rdflib==6.3.2
requests==2.32.3
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
telegram==0.0.1
traits==7.0.2
tzdata==2025.2
urllib3==2.3.0
